refactored directory structure, moved to CMake for builds, and finished CBC mode

+++ test/ecb.cpp
@@ -4,15 +4,16 @@
 
 #include "test.h"
 #include "../lib/aes.h"
-#include "../crypter.h"
+#include "../src/crypter.h"
 
 void Tests::test_ecb(void) {
   uint8_t* data = NULL;
-  uint16_t* data_len = new uint16_t(0);
+  size_t* data_len = new size_t(0);
 
   Crypter::ECB_encrypt(data, data_len, NULL);
   Crypter::ECB_decrypt(data, data_len, NULL);
 
+  // Input data must be a pointer so that we can write new data to it later
   uint8_t* short_data = new uint8_t[4];
   for (int i = 0; i<4; i++) {
     short_data[i] = 0xFF;
